lab template creation structure
